Refactor: Split three modules and shorten a close
rebase and github_client become packages, and github_graphql does too:
neither github_client nor github_graphql was flagged yet, but both sat
one small edit under the file threshold.  close_manager stays a module
and sheds four helpers from _close_single_pr.

should_use_local_rebase took seven parameters.  Its owner, repo and
base branch describe one thing, so they become a BaseRef.

local_rebase_pr is substituted at seven test sites and is called from
within its own package, so it is reached through the module object and
the tests follow it to rebase.local_rebase.  Bypassing that indirection
fails four tests, which is the evidence the substitution still reaches
the code it covers.
