--- src/dependamerge/rebase.py
@@ -1,1089 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: 2026 The Linux Foundation
-"""Rebase strategies for behind pull requests.
-
-This module centralises every code path dependamerge uses to bring
-a PR up to date with its base branch before merging:
-
-- :func:`should_use_local_rebase` decides between the local-git
-  workflow (preserves verified commit signatures via the user's
-  ``~/.gitconfig``) and the GitHub REST ``update-branch``
-  endpoint (fast but produces unsigned commits).
-- :func:`local_rebase_pr` runs the local clone + rebase +
-  force-push-with-lease against a secure temp workspace.
-- :func:`perform_step5_rebase` is the top-level dispatcher used by
-  :class:`AsyncMergeManager._merge_single_pr` Step 5.  It calls
-  :func:`should_use_local_rebase` to decide between paths, then
-  delegates to either ``_run_local_path`` or ``_run_rest_path``
-  (private helpers).  ``_run_rest_path`` runs the REST
-  ``update-branch`` call and the post-rebase polling loop
-  (``_poll_post_rebase``) that waits for GitHub to recompute
-  mergeability.
-- Git network operations authenticate via the ``token`` argument of
-  the :mod:`dependamerge.git_ops` helpers (GIT_ASKPASS-based), so the
-  token never appears in remote URLs, argv, or ``.git/config``.
-
-The dispatcher takes a :class:`RebaseContext` rather than a full
-``AsyncMergeManager`` reference, which keeps the rebase logic
-testable in isolation (no need to construct a manager + GitHub
-client + progress tracker just to exercise a decision tree).
-
-The local-rebase path is the headline reason this module exists:
-``PUT /repos/{owner}/{repo}/pulls/{n}/update-branch`` creates a
-server-side merge commit whose committer is the calling token's
-GitHub user, which is *not* signed with the user's local SSH/GPG
-key.  On repos whose branch protection requires verified
-signatures, the resulting commit loses its ``Verified`` badge and
-becomes un-mergeable.  ``pre-commit-ci[bot]`` PRs are particularly
-affected because that bot has no comment macro for recreating a PR
-with a re-signed commit (https://github.com/pre-commit-ci/issues/issues/41).
-The local path solves this by shelling out to ``git`` so the
-user's signing config is honoured.
-"""
-
-from __future__ import annotations
-
-import asyncio
-import logging
-from collections.abc import Awaitable, Callable
-from dataclasses import dataclass
-from pathlib import Path
-from typing import TYPE_CHECKING, Any
-
-from rich.console import Console
-
-from . import git_ops
-from .bot_identity import is_dependabot
-from .git_ops import (
-    GitError,
-    add_remote,
-    checkout,
-    clone,
-    ensure_git_available,
-    fetch,
-    fetch_branch,
-    push_force_with_lease,
-    rebase,
-    rebase_abort,
-    secure_rmtree,
-)
-from .models import PullRequestInfo
-from .slot_lease import parked
-
-if TYPE_CHECKING:
-    from .github_async import GitHubAsync
-
-
-@dataclass
-class RebaseContext:
-    """Bundle of dependencies the rebase orchestrator needs.
-
-    Passed in lieu of a full ``AsyncMergeManager`` reference so the
-    rebase logic stays decoupled from manager internals and can be
-    tested without standing up the whole merge pipeline.
-    """
-
-    github_client: GitHubAsync | None
-    token: str
-    rebase_local: bool
-    preview_mode: bool
-    merge_recheck_interval: float
-    merge_poll_max_attempts: int
-    log: logging.Logger
-    console: Console
-    # Mutable set on the manager that records PR keys (``owner/repo#N``)
-    # which have already been through Step 5.  Step 5.5 consults this
-    # to avoid doubling the configured ``merge_timeout``.  We keep the
-    # raw set reference rather than a callback so the existing
-    # invariant (Step 5 always adds, Step 5.5 always reads) stays
-    # obvious at the call site.
-    rebased_prs: set[str]
-    # Async callable equivalent to ``manager._enable_auto_merge_for_pr``.
-    # Passed in to avoid a circular import.
-    enable_auto_merge: Callable[[PullRequestInfo, str, str], Awaitable[bool]]
-    # Optional callback (``manager._track_pr_state``) that moves the PR
-    # between transitory states on the Rich progress tracker
-    # ("rebasing", "waiting", or ``None`` to clear).
-    # Default ``None`` keeps existing test constructions working and
-    # makes the tracker strictly optional for isolated use.
-    track_pr_state: Callable[[PullRequestInfo, str | None], None] | None = None
-    # Optional callback (``manager._record_rebase``) that increments the
-    # tracker's cumulative "Rebased" total.  Called once per rebase this
-    # module performs itself — the local force-push path and the REST
-    # ``update-branch`` path.  The ``@dependabot rebase`` macro path is
-    # *not* counted here: ``request_dependabot_rebase`` owns the
-    # accounting for the macro (including its duplicate guard, which
-    # counts nothing), and counting again would double it.
-    record_rebase: Callable[[], None] | None = None
-    # Optional async callable equivalent to
-    # ``manager._request_dependabot_rebase``: posts ``@dependabot
-    # rebase`` on the PR (idempotent — the manager implementation
-    # skips duplicate comments) and returns True when the rebase is
-    # considered requested.  When provided, dependabot PRs that would
-    # otherwise take the local-rebase path use the macro instead:
-    # dependabot force-pushes a freshly *signed* rebase, so there is
-    # no need to clone + rebase + sign locally (which can require
-    # interactive key access, e.g. a YubiKey PIN prompt).  Default
-    # ``None`` preserves the local path for isolated use and existing
-    # tests.
-    request_dependabot_rebase: (
-        Callable[[PullRequestInfo, str, str], Awaitable[bool]] | None
-    ) = None
-
-
-@dataclass
-class Step5Outcome:
-    """Result of :func:`perform_step5_rebase`.
-
-    ``failed`` indicates the caller should mark the PR as ``FAILED``
-    and bail out of the merge attempt.  ``error_message`` is the
-    user-visible reason in that case.
-    """
-
-    failed: bool = False
-    error_message: str | None = None
-
-
-def authed_clone_url(clone_url: str, token: str) -> str:
-    """Deprecated: return ``clone_url`` unchanged.
-
-    Tokens are no longer embedded in clone URLs. Doing so leaked the
-    secret into process listings (``ps`` / ``/proc/<pid>/cmdline``)
-    during the git invocation and persisted it on disk in the
-    workspace's ``.git/config`` (``remote.origin.url``). Authentication
-    is now performed via the ``token`` argument of the
-    :mod:`dependamerge.git_ops` helpers, which supply the secret through
-    a GIT_ASKPASS helper reading from the process environment.
-
-    This shim is retained so third-party callers fail safe (no secret
-    injection) rather than break; it will be removed in a future
-    release.
-    """
-    return clone_url
-
-
-async def should_use_local_rebase(
-    *,
-    github_client: GitHubAsync | None,
-    pr_info: PullRequestInfo,
-    owner: str,
-    repo: str,
-    base_branch: str,
-    rebase_local: bool,
-    log: logging.Logger,
-) -> tuple[bool, str]:
-    """Decide whether Step 5 should rebase locally instead of via REST.
-
-    Returns ``(use_local, reason)``.  ``reason`` is a short
-    human-readable string suitable for debug logging or a
-    user-visible note when ``use_local`` is True.
-
-    The gate activates when ``rebase_local`` is True AND either:
-
-    - the PR is from ``pre-commit-ci[bot]`` (always — that bot has
-      no comment macro for recreating a PR with a re-signed
-      commit), OR
-    - the base branch requires verified signatures AND the current
-      PR head commit is itself verified (so REST update-branch
-      *would* break verification).
-
-    Strict ``is True`` comparison is used on the
-    ``requires_commit_signatures`` return so ``AsyncMock`` defaults
-    don't accidentally route real PRs into the local-rebase path
-    in tests that haven't explicitly set ``return_value``.  If the
-    requirement check raises, we fail safely to the REST path.
-    """
-    if not rebase_local:
-        return False, "--no-rebase-local set"
-
-    # Always use local rebase for pre-commit.ci PRs. The bot has
-    # no comment macro to recover from a verification break, so
-    # we treat it as opt-in regardless of branch protection.
-    if pr_info.author == "pre-commit-ci[bot]":
-        return True, "pre-commit-ci[bot] has no recreate/rebase macro"
-
-    if github_client is None:
-        return False, "no GitHub client"
-
-    # Branch-protection signature requirement (classic + rulesets)
-    try:
-        requires_signatures = await github_client.requires_commit_signatures(
-            owner, repo, base_branch
-        )
-    except Exception as exc:
-        log.debug(
-            "Could not determine signature requirement for %s/%s:%s: %s",
-            owner,
-            repo,
-            base_branch,
-            exc,
-        )
-        return False, "signature requirement check failed"
-
-    # Strict ``is True`` rather than truthy check: ``AsyncMock``
-    # default returns evaluate as truthy, and we explicitly do
-    # not want to enter the network-touching local-rebase path
-    # in test mocks that haven't been set up to handle it.
-    if requires_signatures is not True:
-        return False, "base branch does not require signatures"
-
-    # Base does require verified signatures. Check whether the
-    # current PR head is itself verified — if it isn't, REST
-    # update-branch can't make things worse, so we don't need
-    # the local-rebase machinery.
-    try:
-        all_verified, _unverified = await github_client.check_pr_commit_signatures(
-            owner, repo, pr_info.number
-        )
-    except Exception as exc:
-        log.debug(
-            "Could not check PR commit signatures for %s/%s#%s: %s",
-            owner,
-            repo,
-            pr_info.number,
-            exc,
-        )
-        # Fail closed: if we can't confirm the PR head is
-        # verified, route to the REST path. The opposite
-        # (assuming verification and using the local path)
-        # would mean transient API failures could trigger
-        # network-touching local clones, and would conflict
-        # with the documented gate ("base requires signatures
-        # AND PR head is verified"). When verification isn't
-        # established, REST update-branch can't make things
-        # any worse than they already are.
-        return False, "signature check failed"
-
-    if all_verified:
-        return True, "base requires signatures and PR head is verified"
-    return False, "PR head is not currently verified"
-
-
-async def local_rebase_pr(
-    *,
-    pr_info: PullRequestInfo,
-    owner: str,
-    repo: str,
-    token: str,
-    log: logging.Logger,
-) -> bool:
-    """Rebase a PR locally and force-push the result.
-
-    Clones the head repo into a secure temp workspace, fetches the
-    base branch (from upstream when the PR is from a fork), runs
-    ``git rebase``, and force-pushes with lease back to the head
-    repo.  All git invocations inherit the user's ``~/.gitconfig``,
-    so signing config is respected.
-
-    Returns True only if every step succeeds.  On any failure (no
-    ``git`` on PATH, conflict during rebase, network error, push
-    rejected) the workspace is cleaned up and False is returned;
-    the caller should fall through to the auto-merge path so we
-    never leave a half-applied state.
-    """
-    # Ensure ``git`` is on PATH before we start. ``GitError`` is
-    # also raised when git is missing entirely.
-    try:
-        ensure_git_available()
-    except Exception as exc:
-        log.debug("Local rebase unavailable (no git on PATH?): %s", exc)
-        return False
-
-    # We need the head/base clone URLs. They are populated for PRs
-    # surfaced by recent versions of the find-similar / merge flows;
-    # if missing we synthesise them from the repository names.
-    head_clone_url = pr_info.head_repo_clone_url
-    base_clone_url = pr_info.base_repo_clone_url
-    head_full = pr_info.head_repo_full_name
-    base_full = pr_info.base_repo_full_name or f"{owner}/{repo}"
-
-    # Fail closed when head repo identity is ambiguous. Without a
-    # confirmed ``head_repo_full_name`` or ``head_repo_clone_url``
-    # we cannot tell whether the PR is from a fork. Synthesising a
-    # clone URL from the base repo would push to the wrong remote
-    # for fork PRs (creating or overwriting a branch on the base
-    # repo). The caller falls through to the auto-merge path on
-    # False, which is always safe.
-    if not head_full and not head_clone_url:
-        log.debug(
-            "Local rebase: PR %s/%s#%s missing head_repo identity "
-            "(head_repo_full_name and head_repo_clone_url are both unset); "
-            "failing closed to avoid pushing to the wrong remote.",
-            owner,
-            repo,
-            pr_info.number,
-        )
-        return False
-
-    # Both fields populated, or one of them — synthesise the
-    # missing URL from the known full_name. Safe because
-    # ``head_full`` is now confirmed to refer to the head repo,
-    # not the base.
-    if not head_clone_url:
-        head_clone_url = f"https://github.com/{head_full}.git"
-    if not base_clone_url:
-        base_clone_url = f"https://github.com/{base_full}.git"
-
-    # Decide whether the PR is from a fork *before* we collapse
-    # ``head_full`` to ``base_full`` for clone-URL fallback.
-    # Treating fork PRs as non-fork would cause us to fetch the
-    # base branch from ``origin`` (the fork) instead of
-    # ``upstream`` (the canonical base repo) — either failing
-    # to fetch (the fork doesn't have the latest base commits)
-    # or, worse, fetching stale state and pushing back to the
-    # wrong remote.
-    #
-    # Preference order, all defensive:
-    #   1. The explicit ``pr_info.is_fork`` flag from the API.
-    #   2. Direct comparison of head/base full_names.
-    #   3. Direct comparison of head/base clone URLs.
-    if pr_info.is_fork is not None:
-        is_fork = bool(pr_info.is_fork)
-    elif head_full and base_full and head_full != base_full:
-        is_fork = True
-    elif head_clone_url and base_clone_url and head_clone_url != base_clone_url:
-        is_fork = True
-    else:
-        is_fork = False
-
-    # Now safe to collapse ``head_full`` for the clone-URL
-    # synthesis fallback. ``is_fork`` has already been computed
-    # above and won't be misled by this assignment.
-    head_full = head_full or base_full
-
-    head_branch = pr_info.head_branch
-    base_branch = pr_info.base_branch or "main"
-    if not head_branch:
-        log.debug(
-            "Local rebase: PR %s/%s#%s missing head_branch",
-            owner,
-            repo,
-            pr_info.number,
-        )
-        return False
-
-    origin_url = head_clone_url
-    upstream_url = base_clone_url
-
-    # Use a per-PR workspace under a secure temp parent so
-    # concurrent rebases (--concurrency=N) don't collide.
-    workspace_parent = Path(
-        git_ops.create_secure_tempdir(prefix="dependamerge-localrebase-")
-    )
-    workspace = (
-        workspace_parent
-        / f"{(head_full or base_full).replace('/', '__')}__pr_{pr_info.number}"
-    )
-    workspace.mkdir(parents=True, exist_ok=True)
-
-    try:
-        # Clone the head repo at the PR's head branch. Shallow
-        # clone keeps disk + network footprint low for what are
-        # typically tiny dependency-update PRs.
-        try:
-            clone(
-                origin_url,
-                workspace,
-                branch=head_branch,
-                depth=50,
-                single_branch=True,
-                no_tags=True,
-                filter_blobs=True,
-                logger=log.debug,
-                token=token,
-            )
-        except GitError as exc:
-            log.debug("Local rebase: clone failed for %s: %s", pr_info.html_url, exc)
-            return False
-
-        # Fetch the base branch — from upstream when the PR is
-        # from a fork, from origin otherwise. We need it
-        # available locally before we can rebase onto it.
-        #
-        # Use :func:`fetch_branch` rather than the bare ``fetch``
-        # form: the ``--single-branch`` clone above restricts the
-        # remote's configured refspec to the PR head branch, so a
-        # bare ``git fetch origin <base>`` would only populate
-        # ``FETCH_HEAD`` and a subsequent ``git rebase origin/<base>``
-        # would fail with ``fatal: invalid upstream 'origin/<base>'``.
-        try:
-            if is_fork:
-                add_remote("upstream", upstream_url, cwd=workspace, logger=log.debug)
-                fetch_branch(
-                    "upstream",
-                    base_branch,
-                    cwd=workspace,
-                    depth=50,
-                    logger=log.debug,
-                    token=token,
-                )
-                rebase_onto = f"upstream/{base_branch}"
-            else:
-                fetch_branch(
-                    "origin",
-                    base_branch,
-                    cwd=workspace,
-                    depth=50,
-                    logger=log.debug,
-                    token=token,
-                )
-                rebase_onto = f"origin/{base_branch}"
-        except GitError as exc:
-            log.debug("Local rebase: fetch failed for %s: %s", pr_info.html_url, exc)
-            return False
-
-        # Make sure we are on the head branch (defensive against
-        # detached HEAD after clone --branch).
-        try:
-            checkout(head_branch, cwd=workspace, create=False, logger=log.debug)
-        except GitError:
-            # Already on the branch, or branch missing locally;
-            # rebase will surface the real problem if any.
-            pass
-
-        # Rebase. ``git rebase`` runs with ``check=False`` (see
-        # ``git_ops.rebase``), so a non-zero exit does *not* raise
-        # ``GitError``; we have to inspect ``returncode``
-        # ourselves. Conflicts are the most common cause of a
-        # non-zero exit here, but other failures (corrupt index,
-        # invalid base ref, etc.) hit the same path — surface
-        # stderr/stdout in debug output so the cause is visible
-        # to anyone investigating, then abort the rebase to leave
-        # the workspace in a clean state before cleanup.
-        rebase_result = rebase(
-            rebase_onto,
-            cwd=workspace,
-            autostash=False,
-            interactive=False,
-            logger=log.debug,
-        )
-
-        if rebase_result.returncode != 0:
-            # Shallow clones (depth=50) can miss the merge base
-            # for PRs whose branch point is older than 50 commits.
-            # ``git rebase`` reports this as a generic non-zero
-            # exit; the diagnostic is visible in stderr but we
-            # can't reliably distinguish it without parsing
-            # locale-dependent text.
-            #
-            # Recovery: abort the failed rebase, unshallow both
-            # remotes, and retry the rebase. If that also fails,
-            # the cause is genuine (conflicts, corrupt index,
-            # etc.) and we return False as before so the caller
-            # falls through to the auto-merge path.
-            log.debug(
-                "Local rebase: rebase exited non-zero for %s "
-                "(rc=%d, stderr=%r, stdout=%r); attempting "
-                "unshallow + retry.",
-                pr_info.html_url,
-                rebase_result.returncode,
-                rebase_result.stderr,
-                rebase_result.stdout,
-            )
-            try:
-                rebase_abort(cwd=workspace, logger=log.debug)
-            except Exception:
-                # Cleanup abort is best-effort; ignore so the original
-                # rebase failure still drives the retry below.
-                pass
-
-            try:
-                fetch(
-                    "origin",
-                    cwd=workspace,
-                    unshallow=True,
-                    logger=log.debug,
-                    token=token,
-                )
-                if is_fork:
-                    fetch(
-                        "upstream",
-                        cwd=workspace,
-                        unshallow=True,
-                        logger=log.debug,
-                        token=token,
-                    )
-            except GitError as exc:
-                log.debug(
-                    "Local rebase: unshallow failed for %s: %s",
-                    pr_info.html_url,
-                    exc,
-                )
-                return False
-
-            rebase_result = rebase(
-                rebase_onto,
-                cwd=workspace,
-                autostash=False,
-                interactive=False,
-                logger=log.debug,
-            )
-            if rebase_result.returncode != 0:
-                log.debug(
-                    "Local rebase: rebase still failing after unshallow "
-                    "for %s (rc=%d, stderr=%r); aborting.",
-                    pr_info.html_url,
-                    rebase_result.returncode,
-                    rebase_result.stderr,
-                )
-                try:
-                    rebase_abort(cwd=workspace, logger=log.debug)
-                except Exception:
-                    # Cleanup abort is best-effort; ignore and return
-                    # failure to the caller regardless.
-                    pass
-                return False
-            log.debug(
-                "Local rebase: succeeded after unshallow for %s",
-                pr_info.html_url,
-            )
-
-        # Force-push with lease to the head repo. We push back to
-        # ``origin`` because the head ref always lives there (even
-        # for forks, the head repo *is* the fork).
-        try:
-            push_force_with_lease(
-                "origin",
-                head_branch,
-                head_branch,
-                cwd=workspace,
-                logger=log.debug,
-                token=token,
-            )
-        except GitError as exc:
-            log.debug(
-                "Local rebase: force-push failed for %s: %s",
-                pr_info.html_url,
-                exc,
-            )
-            return False
-
-        log.debug("Local rebase succeeded for %s", pr_info.html_url)
-        return True
-
-    finally:
-        # Always clean up. The workspace contains a clone of the
-        # user's repository, so we want it gone even on success.
-        try:
-            secure_rmtree(workspace_parent)
-        except Exception as exc:
-            log.debug(
-                "Local rebase: failed to clean up workspace %s: %s",
-                workspace_parent,
-                exc,
-            )
-
-
-async def perform_step5_rebase(
-    *,
-    ctx: RebaseContext,
-    pr_info: PullRequestInfo,
-    owner: str,
-    repo: str,
-) -> Step5Outcome:
-    """Run Step 5 of the merge flow: bring the PR up to date with its base.
-
-    Dispatches between the local-git path (signature-preserving)
-    and the legacy REST ``update-branch`` path based on
-    :func:`should_use_local_rebase`.  When the local path is
-    selected, REST ``update-branch`` is **never** called — even on
-    local-rebase failure — so we never destroy a verified
-    signature.  In the failure case we mark the PR as having been
-    through Step 5 (so Step 5.5 doesn't double the configured
-    ``merge_timeout``) and let auto-merge take over server-side.
-
-    Returns a :class:`Step5Outcome`.  ``failed=True`` indicates the
-    caller should set ``MergeStatus.FAILED`` and bail; the legacy
-    REST path is the only path that can produce this outcome (a
-    raised exception during ``update_branch`` or the polling loop).
-    """
-    if ctx.preview_mode:
-        # NOTE: In preview mode, we should NOT print here as it
-        # breaks single-line reporting.  The preview output
-        # should only be a single line per PR in the evaluation
-        # section.
-        return Step5Outcome()
-
-    ctx.log.debug("Rebasing %s [behind base branch]", pr_info.html_url)
-    _set_tracker_state(ctx, pr_info, "rebasing")
-
-    use_local, local_reason = await should_use_local_rebase(
-        github_client=ctx.github_client,
-        pr_info=pr_info,
-        owner=owner,
-        repo=repo,
-        base_branch=pr_info.base_branch or "main",
-        rebase_local=ctx.rebase_local,
-        log=ctx.log,
-    )
-
-    if use_local:
-        # For dependabot PRs, prefer the ``@dependabot rebase`` comment
-        # macro over a local clone + rebase + sign: dependabot
-        # force-pushes a freshly signed rebase itself, which satisfies
-        # the very signature requirement that routed us to the local
-        # path — without touching the operator's signing key (a local
-        # rebase on a signature-requiring repo can trigger interactive
-        # prompts, e.g. a YubiKey PIN).  Falls back to the local path
-        # when the macro cannot be posted.
-        if is_dependabot(pr_info.author) and ctx.request_dependabot_rebase is not None:
-            handled = await _run_dependabot_macro_path(
-                ctx=ctx,
-                pr_info=pr_info,
-                owner=owner,
-                repo=repo,
-                local_reason=local_reason,
-            )
-            if handled:
-                return Step5Outcome()
-        await _run_local_path(
-            ctx=ctx,
-            pr_info=pr_info,
-            owner=owner,
-            repo=repo,
-            local_reason=local_reason,
-        )
-        return Step5Outcome()
-
-    return await _run_rest_path(
-        ctx=ctx,
-        pr_info=pr_info,
-        owner=owner,
-        repo=repo,
-    )
-
-
-def _set_tracker_state(
-    ctx: RebaseContext,
-    pr_info: PullRequestInfo,
-    state: str | None,
-) -> None:
-    """Move the PR between transitory progress-tracker states.
-
-    No-op when the context was constructed without a
-    ``track_pr_state`` callback (isolated tests, preview mode).
-    """
-    if ctx.track_pr_state is not None:
-        ctx.track_pr_state(pr_info, state)
-
-
-def _record_rebase(ctx: RebaseContext) -> None:
-    """Count one rebase operation on the progress tracker.
-
-    No-op when the context was constructed without a ``record_rebase``
-    callback (isolated tests).  Preview runs never reach here at all:
-    :func:`perform_step5_rebase` returns before dispatching to any
-    rebase path.
-    """
-    if ctx.record_rebase is not None:
-        ctx.record_rebase()
-
-
-async def _run_local_path(
-    *,
-    ctx: RebaseContext,
-    pr_info: PullRequestInfo,
-    owner: str,
-    repo: str,
-    local_reason: str,
-) -> None:
-    """Local-rebase path.  Always succeeds from the caller's POV.
-
-    Whether the underlying ``git`` workflow succeeds or fails, we
-    never fall back to REST ``update-branch`` (doing so would
-    defeat the whole point of the local path).
-
-    Auto-merge enablement and ``_rebased_prs`` marking are linked:
-
-    - If auto-merge gets enabled, mark ``_rebased_prs`` so Step
-      5.5 skips this PR; auto-merge will handle the bounded wait
-      server-side and Step 6's skip gate routes to
-      ``AUTO_MERGE_PENDING``.
-    - If auto-merge cannot be enabled (repo doesn't allow it,
-      branch protection blocks it, etc.), do **not** mark
-      ``_rebased_prs`` so Step 5.5 still runs its bounded poll
-      loop. Without this, GitHub may still be recomputing
-      mergeability after the force-push when Step 6 fires and a
-      manual merge attempt would 405 transiently. Letting Step
-      5.5 wait gives GitHub time to settle before Step 6 acts.
-    """
-    ctx.log.debug(
-        "Local rebase: %s [%s]",
-        pr_info.html_url,
-        local_reason,
-    )
-    try:
-        local_rebase_ok = await local_rebase_pr(
-            pr_info=pr_info,
-            owner=owner,
-            repo=repo,
-            token=ctx.token,
-            log=ctx.log,
-        )
-    except Exception as exc:
-        ctx.log.debug(
-            "Local rebase raised unexpectedly for %s: %s",
-            pr_info.html_url,
-            exc,
-        )
-        local_rebase_ok = False
-
-    # Try to enable auto-merge. We capture the return value so we
-    # can decide whether Step 5.5 should also run (see comment
-    # below).
-    try:
-        auto_merge_ok = await ctx.enable_auto_merge(pr_info, owner, repo)
-    except Exception as exc:
-        ctx.log.debug(
-            "Could not enable auto-merge after local rebase for %s: %s",
-            pr_info.html_url,
-            exc,
-        )
-        auto_merge_ok = False
-
-    # Only mark ``_rebased_prs`` when auto-merge is active.
-    # Otherwise leave it unset so Step 5.5 still runs its bounded
-    # poll loop — GitHub may still be recomputing mergeability
-    # after the force-push, and a manual merge attempt in Step 6
-    # without that wait would 405 transiently.  When auto-merge
-    # *is* active, Step 6's skip gate routes the PR to
-    # ``AUTO_MERGE_PENDING`` directly, so the Step 5.5 wait would
-    # only double the merge_timeout.
-    if auto_merge_ok:
-        ctx.rebased_prs.add(f"{owner}/{repo}#{pr_info.number}")
-
-    # Either way the rebase attempt is over, so clear the "rebasing"
-    # state ``perform_step5_rebase`` set before dispatch.  Leaving it
-    # set on the failure branch would strand the PR displaying as
-    # "Rebasing" for the rest of the run, since this path defers to
-    # auto-merge rather than reaching a terminal outcome that would
-    # clear it.
-    _set_tracker_state(ctx, pr_info, None)
-    if local_rebase_ok:
-        ctx.log.debug("Rebased (local): %s", pr_info.html_url)
-        # The cumulative "Rebased" total is what keeps a record of the
-        # rebase from here on.
-        _record_rebase(ctx)
-    else:
-        ctx.log.debug(
-            "Local rebase failed; deferring to auto-merge: %s",
-            pr_info.html_url,
-        )
-
-
-async def _run_dependabot_macro_path(
-    *,
-    ctx: RebaseContext,
-    pr_info: PullRequestInfo,
-    owner: str,
-    repo: str,
-    local_reason: str,
-) -> bool:
-    """Request a rebase via the ``@dependabot rebase`` macro.
-
-    Used instead of the local-rebase path for dependabot PRs: the bot
-    rebases the branch onto the current base and force-pushes a
-    commit signed with its own key, preserving the ``Verified`` badge
-    that signature-requiring branch protection demands.
-
-    The rebase completes asynchronously (dependabot typically takes
-    one to a few minutes), so this path never waits for it.  It mirrors
-    the local path's auto-merge contract instead:
-
-    - Auto-merge armed → mark ``_rebased_prs`` so Step 5.5 skips the
-      PR and Step 6's skip gate routes it to ``AUTO_MERGE_PENDING``;
-      GitHub merges server-side once the rebase lands and checks pass.
-    - Auto-merge unavailable → leave the PR unmarked so Step 5.5
-      still runs its bounded wait for the rebase + checks.
-
-    Returns True when the macro was posted (or was already pending);
-    False when it could not be requested — the caller then falls back
-    to the local-rebase path.
-    """
-    if ctx.request_dependabot_rebase is None:
-        return False
-
-    ctx.log.debug(
-        "Dependabot rebase macro: %s [%s]",
-        pr_info.html_url,
-        local_reason,
-    )
-    try:
-        requested = await ctx.request_dependabot_rebase(pr_info, owner, repo)
-    except Exception as exc:
-        ctx.log.debug(
-            "Dependabot rebase request raised for %s: %s",
-            pr_info.html_url,
-            exc,
-        )
-        requested = False
-    if not requested:
-        return False
-
-    try:
-        auto_merge_ok = await ctx.enable_auto_merge(pr_info, owner, repo)
-    except Exception as exc:
-        ctx.log.debug(
-            "Could not enable auto-merge after dependabot rebase request for %s: %s",
-            pr_info.html_url,
-            exc,
-        )
-        auto_merge_ok = False
-
-    # Same marking rule as the local path: only skip Step 5.5 when
-    # auto-merge is armed to finish the job server-side.
-    if auto_merge_ok:
-        ctx.rebased_prs.add(f"{owner}/{repo}#{pr_info.number}")
-
-    ctx.log.debug("Rebase requested (dependabot macro): %s", pr_info.html_url)
-    # No counting here: ``ctx.request_dependabot_rebase`` owns the
-    # cumulative totals and records them only when it actually posts
-    # the macro (its duplicate guard returns True without posting when
-    # an earlier run already requested the rebase).  Counting again
-    # here would double the posted case and invent the guarded one.
-    _set_tracker_state(ctx, pr_info, None)
-    return True
-
-
-async def _run_rest_path(
-    *,
-    ctx: RebaseContext,
-    pr_info: PullRequestInfo,
-    owner: str,
-    repo: str,
-) -> Step5Outcome:
-    """Legacy REST ``update-branch`` path with post-rebase polling.
-
-    Uses the GitHub REST API to bring the PR up to date, enables
-    auto-merge so the PR merges even if we time out waiting for
-    status checks, then polls until checks complete or
-    ``merge_timeout`` elapses.  Updates ``pr_info`` in place with
-    the post-rebase state.
-
-    Returns a :class:`Step5Outcome` whose ``failed`` field is True
-    when ``update_branch`` (or the polling apparatus) raises an
-    exception — the caller should mark the merge as ``FAILED`` in
-    that case.
-    """
-    if ctx.github_client is None:
-        return Step5Outcome(failed=True, error_message="GitHub client not initialized")
-    client = ctx.github_client
-
-    try:
-        await client.update_branch(owner, repo, pr_info.number)
-        _record_rebase(ctx)
-
-        # Enable auto-merge so the PR merges even if we time out
-        # waiting for status checks.
-        auto_merge_ok = await ctx.enable_auto_merge(pr_info, owner, repo)
-        if auto_merge_ok:
-            ctx.log.debug(
-                "Auto-merge enabled after rebase for %s/%s#%s",
-                owner,
-                repo,
-                pr_info.number,
-            )
-
-        # Wait briefly for GitHub to start processing the update.
-        # The full recheck interval (default 10s) is unnecessary here:
-        # ``_poll_post_rebase`` polls at that cadence anyway and
-        # tolerates the transient ``null``/``behind`` states GitHub
-        # reports while recomputing, so a short head start just gets
-        # the first data point sooner.  The settle sleep and the poll
-        # are both waits on GitHub-side processing, so the worker's
-        # concurrency slot is released for their duration
-        # (``parked()`` — see ``slot_lease.py``).
-        ctx.log.debug("Waiting for rebase to process: %s", pr_info.html_url)
-        _set_tracker_state(ctx, pr_info, "waiting")
-        async with parked():
-            await asyncio.sleep(min(2.0, ctx.merge_recheck_interval))
-
-            updated_mergeable, updated_mergeable_state = await _poll_post_rebase(
-                ctx=ctx,
-                pr_info=pr_info,
-                owner=owner,
-                repo=repo,
-                auto_merge_ok=auto_merge_ok,
-            )
-
-        # Update our PR info with the latest state.  Preserve the
-        # previous non-None values when the refresh returns
-        # ``null`` (GitHub is still computing).  The Step 6
-        # auto-merge skip gate accepts both ``True`` and ``None``
-        # (it excludes only the explicit ``False`` case), so a
-        # transient null no longer blocks the auto-merge path on
-        # its own.  We still preserve the prior known ``True`` so
-        # downstream logging and any future tightening of that
-        # predicate get an accurate state to work with.  The same
-        # rationale applies to ``mergeable_state``: GitHub returns
-        # ``null`` while still computing, and the post-rebase
-        # reporting / Step 5.5 logic branches on this value (e.g.
-        # "clean" vs "blocked" vs "behind"); a transient ``None``
-        # would otherwise be classified as the catch-all "other
-        # state" branch.
-        if updated_mergeable is not None:
-            pr_info.mergeable = updated_mergeable
-        if updated_mergeable_state is not None:
-            pr_info.mergeable_state = updated_mergeable_state
-
-        # Mark this PR as having gone through the Step 5 rebase
-        # + poll path.  Step 5.5 will consult ``_rebased_prs`` to
-        # avoid doubling the merge_timeout when the rebase exits
-        # in ``blocked`` or ``behind`` state.
-        ctx.rebased_prs.add(f"{owner}/{repo}#{pr_info.number}")
-
-        _set_tracker_state(ctx, pr_info, None)
-        _log_post_rebase_status(ctx=ctx, pr_info=pr_info)
-        return Step5Outcome()
-
-    except Exception as exc:
-        ctx.log.warning("Rebase failed for %s: %s", pr_info.html_url, exc)
-        return Step5Outcome(failed=True, error_message=f"Failed to rebase PR: {exc}")
-
-
-async def _poll_post_rebase(
-    *,
-    ctx: RebaseContext,
-    pr_info: PullRequestInfo,
-    owner: str,
-    repo: str,
-    auto_merge_ok: bool,
-) -> tuple[bool | None, str | None]:
-    """Poll the PR after ``update_branch`` until it stabilises.
-
-    Returns the latest ``(mergeable, mergeable_state)`` observed.
-    Updates ``pr_info.head_sha`` in place when the refresh shows a
-    new head commit (so any subsequent ``analyze_block_reason()``
-    call queries the rebased commit, not the pre-rebase one).
-    """
-    if ctx.github_client is None:
-        return pr_info.mergeable, pr_info.mergeable_state
-    client = ctx.github_client
-
-    updated_mergeable: bool | None = pr_info.mergeable
-    updated_mergeable_state: str | None = pr_info.mergeable_state
-
-    for check_attempt in range(ctx.merge_poll_max_attempts):
-        updated_pr_data: Any = await client.get(
-            f"/repos/{owner}/{repo}/pulls/{pr_info.number}"
-        )
-
-        if isinstance(updated_pr_data, dict):
-            updated_mergeable = updated_pr_data.get("mergeable")
-            updated_mergeable_state = updated_pr_data.get("mergeable_state")
-            updated_head = (updated_pr_data.get("head") or {}).get("sha")
-            if updated_head:
-                pr_info.head_sha = updated_head
-        else:
-            updated_mergeable = None
-            updated_mergeable_state = None
-
-        if _poll_should_continue(
-            ctx=ctx,
-            pr_info=pr_info,
-            attempt=check_attempt,
-            mergeable_state=updated_mergeable_state,
-            auto_merge_ok=auto_merge_ok,
-        ):
-            await asyncio.sleep(ctx.merge_recheck_interval)
-            continue
-        break
-
-    return updated_mergeable, updated_mergeable_state
-
-
-def _poll_should_continue(
-    *,
-    ctx: RebaseContext,
-    pr_info: PullRequestInfo,
-    attempt: int,
-    mergeable_state: str | None,
-    auto_merge_ok: bool,
-) -> bool:
-    """Return True when the post-rebase poll loop should keep waiting.
-
-    Centralising the per-state decisions here keeps
-    :func:`_poll_post_rebase` short and readable.
-    """
-    if mergeable_state == "clean":
-        return False
-
-    last_attempt = attempt >= ctx.merge_poll_max_attempts - 1
-
-    if mergeable_state == "behind":
-        if last_attempt:
-            return False
-        ctx.log.debug(
-            "PR still processing rebase, waiting... (attempt %d/%d)",
-            attempt + 1,
-            ctx.merge_poll_max_attempts,
-        )
-        return True
-
-    if mergeable_state == "blocked":
-        if last_attempt:
-            _log_blocked_timeout(ctx=ctx, pr_info=pr_info, auto_merge_ok=auto_merge_ok)
-            return False
-        ctx.log.debug(
-            "PR status checks running after rebase, waiting... (attempt %d/%d)",
-            attempt + 1,
-            ctx.merge_poll_max_attempts,
-        )
-        return True
-
-    if mergeable_state is None:
-        # GitHub is still computing mergeability (typically right
-        # after update_branch).  Treat as transient and keep
-        # polling until the deadline or a concrete state arrives —
-        # breaking here would otherwise exit prematurely and (if
-        # auto-merge enablement failed) fall through to a manual
-        # merge attempt against the still-resolving PR state.
-        if last_attempt:
-            return False
-        ctx.log.debug(
-            "PR mergeable_state still computing after rebase, "
-            "waiting... (attempt %d/%d)",
-            attempt + 1,
-            ctx.merge_poll_max_attempts,
-        )
-        return True
-
-    # Any other concrete state ("dirty", "draft", "unstable",
-    # "unknown", ...) ends the poll loop immediately.
-    return False
-
-
-def _log_blocked_timeout(
-    *,
-    ctx: RebaseContext,
-    pr_info: PullRequestInfo,
-    auto_merge_ok: bool,
-) -> None:
-    """Log when the post-rebase poll times out blocked (log only)."""
-    if auto_merge_ok:
-        ctx.log.warning(
-            "Auto-merge will complete: %s [timeout waiting for checks]",
-            pr_info.html_url,
-        )
-    else:
-        ctx.log.warning(
-            "Proceeding without checks: %s [timeout waiting for checks]",
-            pr_info.html_url,
-        )
-
-
-def _log_post_rebase_status(
-    *,
-    ctx: RebaseContext,
-    pr_info: PullRequestInfo,
-) -> None:
-    """Log the post-rebase status based on the final mergeable_state."""
-    state = pr_info.mergeable_state
-    if state == "clean":
-        ctx.log.debug("Rebased: %s", pr_info.html_url)
-    elif state == "behind":
-        ctx.log.debug("Rebased: %s [still behind after rebase]", pr_info.html_url)
-    elif state == "blocked":
-        ctx.log.debug("Rebased: %s [waiting for status checks]", pr_info.html_url)
-    else:
-        ctx.log.debug("Rebased: %s [state=%s]", pr_info.html_url, state)
